Two Changes
-Changes text to ‘Movies out soon” in html for upcoming movies page

-Changes links to the API for both the out now and upcoming pages. The
upcoming link I think is worth keeping.

The out now page I changed the RT link to Opening. I’m not sure if RT
changes the list on Monday or on Wed/Thur. So, for example,
Interstellar is on the Opening API list but not in the Out In Theaters
list.

**Maybe the solution will mean we have to make a third page. One for
Movies Opening, One for Movies already Out In Theaters, and One for
Upcoming Movies.**